[Refactor] 스크린 분리
