+++ app/src/main/java/org/sopt/at/home/BarNavHost.kt
@@ -5,16 +5,14 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import org.sopt.at.home.screen.History
-import org.sopt.at.home.screen.HistoryScreen
-import org.sopt.at.home.screen.Home
-import org.sopt.at.home.screen.HomeScreen
-import org.sopt.at.home.screen.Live
-import org.sopt.at.home.screen.LiveScreen
-import org.sopt.at.home.screen.Search
-import org.sopt.at.home.screen.SearchScreen
-import org.sopt.at.home.screen.Shorts
-import org.sopt.at.home.screen.ShortsScreen
+import org.sopt.at.history.History
+import org.sopt.at.history.HistoryScreen
+import org.sopt.at.live.Live
+import org.sopt.at.live.LiveScreen
+import org.sopt.at.search.Search
+import org.sopt.at.search.SearchScreen
+import org.sopt.at.shorts.Shorts
+import org.sopt.at.shorts.ShortsScreen
 import org.sopt.at.my.My
 import org.sopt.at.my.MyScreen
