+++ app/src/main/java/org/sopt/at/live/LiveScreen.kt
@@ -1,4 +1,4 @@
-package org.sopt.at.home.screen
+package org.sopt.at.live
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
